add standard mod project, including automatic references & mod deploy
