+++ .gitignore
@@ -1,7 +1,25 @@
-*.csproj
+# user-specific files
+*.suo
 *.user
-packages.config
-bin/
-obj/
-Properties/
-Service References/
+*.userosscache
+*.sln.docstates
+
+# build results
+[Dd]ebug/
+[Rr]elease/
+[Bb]in/
+[Oo]bj/
+
+# Visual Studio cache/options
+.vs/
+
+# ReSharper
+_ReSharper*/
+*.[Rr]e[Ss]harper
+*.DotSettings.user
+
+# NuGet packages
+*.nupkg
+**/packages/*
+*.nuget.props
+*.nuget.targets
