+++ PartOfTheCommunity.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26403.7
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PartOfTheCommunity", "PartOfTheCommunity\PartOfTheCommunity.csproj", "{58913321-94A9-4B11-95C4-FABAEF78AE76}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".metadata", ".metadata", "{6853CA34-AFA2-4237-90FB-18F2BF5BB910}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		README.md = README.md
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{58913321-94A9-4B11-95C4-FABAEF78AE76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{58913321-94A9-4B11-95C4-FABAEF78AE76}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{58913321-94A9-4B11-95C4-FABAEF78AE76}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{58913321-94A9-4B11-95C4-FABAEF78AE76}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
